added support for optional content type in command
This change makes it possible to use the assistant-notifier plugin to stream any kind of media. When a content type is specified, the client is disconnected without stopping in order to support long-running media (e.g., video streaming) without keeping an active client in memory. Consequently, no notification is received when playing ends, and the volume would not be reverted to its previous setting, which is why volume is not changed when a content type argument is specified.